requirements.txt
locustio==0.7.5
pyquery==1.2.17

